rename src/helpers -> utils
fix signOut
