+++ src/components/Artist/albums.tsx
@@ -3,7 +3,7 @@ import { ItemCard } from '../Global/components/item-card'
 import { ArtistAlbumsProps, ArtistEpsProps, ArtistFeaturedOnProps } from '../types'
 import { Text } from '../Global/helpers/text'
 import { useArtistContext } from '../../providers/Artist'
-import { convertRunTimeTicksToSeconds } from '../../helpers/runtimeticks'
+import { convertRunTimeTicksToSeconds } from '../../utils/runtimeticks'
 import Animated, { useAnimatedScrollHandler } from 'react-native-reanimated'
 import { ActivityIndicator } from 'react-native'
 import { useSafeAreaFrame } from 'react-native-safe-area-context'
