+++ src/components/Login/screens/server-address.tsx
@@ -17,6 +17,7 @@ import Icon from '../../Global/components/icon'
 import { PublicSystemInfo } from '@jellyfin/sdk/lib/generated-client/models'
 import { connectToServer } from '../../../api/mutations/login'
 import { IS_MAESTRO_BUILD } from '../../../configs/config'
+import { sleepify } from '../../../utils/sleep'
 
 export default function ServerAddress({
 	navigation,
@@ -43,7 +44,7 @@ export default function ServerAddress({
 	}, [serverAddress])
 
 	useEffect(() => {
-		signOut()
+		sleepify(250).then(() => signOut())
 	}, [])
 
 	const useServerMutation = useMutation({
