+++ src/player/handlers.ts
@@ -1,7 +1,7 @@
 import { Progress, State } from 'react-native-track-player'
 import JellifyTrack from '../types/JellifyTrack'
 import { PlaystateApi } from '@jellyfin/sdk/lib/generated-client/api/playstate-api'
-import { convertSecondsToRunTimeTicks } from '../helpers/runtimeticks'
+import { convertSecondsToRunTimeTicks } from '../utils/runtimeticks'
 
 export async function handlePlaybackState(
 	sessionId: string,
